Moved timeConversion into its own file; still using DocBrown command
